+++ src/ItemsPerHoldingsRecord.js
@@ -107,8 +107,8 @@ class ItemsPerHoldingsRecord extends React.Component {
 
   renderButtonsGroup = () => {
     const {
-      instance,
       holdingsRecord,
+      location: { pathname },
     } = this.props;
 
     return (
@@ -116,7 +116,7 @@ class ItemsPerHoldingsRecord extends React.Component {
         <Button
           id="clickable-view-holdings"
           data-test-view-holdings
-          to={{ pathname: `/inventory/view/${instance.id}/${holdingsRecord.id}` }}
+          to={{ pathname: `${pathname}/${holdingsRecord.id}` }}
           style={{ marginRight: '5px' }}
         >
           <FormattedMessage id="ui-inventory.viewHoldings" />
@@ -232,6 +232,7 @@ class ItemsPerHoldingsRecord extends React.Component {
 
 ItemsPerHoldingsRecord.propTypes = {
   instance: PropTypes.object,
+  location: PropTypes.object,
   holdingsRecord: PropTypes.object.isRequired,
   mutator: PropTypes.shape({
     items: PropTypes.shape({